fastapi_service/requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==24.2.0
click==8.1.7
fastapi==0.115.0
frozenlist==1.4.1
grpcio==1.66.1
grpcio-tools==1.66.1
h11==0.14.0
idna==3.10
multidict==6.1.0
pika==1.3.2
protobuf==5.28.2
psycopg2-binary==2.9.9
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
python-logstash==0.4.8
sniffio==1.3.1
starlette==0.38.5
typing_extensions==4.12.2
uvicorn==0.30.6
yarl==1.11.1
